P0-1: Promote Gen0Count/Gen1Count/Gen2Count to long to prevent overflow corruption
- Change Gen0Count/Gen1Count/Gen2Count in TypeAggregateIndexEntry from int (4B each) to long (8B each)
- Bump TypeAggregateIndex binary format version from 2 to 3
- Increase TypeEntrySize from 68 to 80 bytes (gain 12 bytes for 3×int→long promotion)
- Update WriteTypeEntry and ReadTypeEntry with new byte offsets for 8-byte gen count fields
- Update all consumer code: TypeGenerationProfile, GCGenerationDomainResult,
  FinalizableObjectDomainResult, StringDomainResult to use long for gen counts
- Update analyzers and InsightEngine to handle long gen count values
- Fixes silent data corruption on large heaps where per-type object counts exceed 2.1B

This is a critical fix preventing int overflow corruption when processing dumps with
billions of objects per type. The cast of negative int values to ulong was producing
massive ~18EB estimates.

+++ src/DumpDetective.Analysis/Analyzers/DominatorAnalyzer.cs
@@ -169,7 +169,7 @@ public sealed class DominatorAnalyzer : IAnalyzer, IParallelHeapIndexScanPartici
         if (cache is HeapAnalysisCache heapCache && heapCache.TryGetHeapIndex(out HeapIndexBuildResult? heapIndex))
             aggregates = heapIndex.TypeAggregates;
 
-        var candidates = new List<(string TypeName, ulong SampleAddress, int Count, ulong TotalSize, ulong LohSize, int Gen2Count, ulong Score)>(capacity: Math.Min(32, typeStats.Count));
+        var candidates = new List<(string TypeName, ulong SampleAddress, int Count, ulong TotalSize, ulong LohSize, long Gen2Count, ulong Score)>(capacity: Math.Min(32, typeStats.Count));
 
         foreach (KeyValuePair<string, CachedTypeStatistics> kv in typeStats)
         {
@@ -182,7 +182,7 @@ public sealed class DominatorAnalyzer : IAnalyzer, IParallelHeapIndexScanPartici
             ulong totalSize = kv.Value.TotalSize;
             ulong lohSize = kv.Value.LohSize;
             int count = kv.Value.Count;
-            int gen2Count = 0;
+            long gen2Count = 0;
 
             if (aggregates is not null)
             {
